No longer asks for password and works for captial search terms
Instead having to trust us with logins and passwords, users now
just choose if the current run should be personalized.  If they
do, then the bot with drop them off at the sign-in page and
Google can handle the login and 2fac.  The bot justs waits
until they're back at the search page.

The bot wouldn't proceed past searches beginning with capital
letters before because it converted the page title to lower
case.  Now it doesn't.